cleaned up the product api call
