+++ blocks/product-children/product-children.js
@@ -87,7 +87,7 @@ const childProducts = `
 export default async function decorate(block) {
   const sku = getSKU();
   const host = (window.location.host === 'lifesciences.danaher.com') ? window.location.host : 'stage.lifesciences.danaher.com';
-  const response = getProductResponse();
+  const response = await getProductResponse();
   if (response?.length > 0 && response[0]?.raw?.objecttype === 'Family' && response[0]?.raw?.numproducts > 0) {
     block.classList.add('pt-10');
     block.innerHTML = childProducts;
