+++ blocks/product-hero/product-hero.js
@@ -158,7 +158,7 @@ async function addToQuote(product) {
 }
 
 export default async function decorate(block) {
-  const response = getProductResponse();
+  const response = await getProductResponse();
   if (response?.length > 0) {
     document.title = response[0].Title ? response[0].Title : 'Danaher Product';
     const allImages = response[0]?.raw.images;
