+++ blocks/product-overview/product-overview.js
@@ -1,7 +1,7 @@
 import { getProductResponse } from '../../scripts/scripts.js';
 
 export default async function decorate(block) {
-  const response = getProductResponse();
+  const response = await getProductResponse();
   if (response?.length > 0) {
     block.innerHTML = response[0].raw?.richlongdescription ? response[0].raw.richlongdescription : '';
     block.classList.add(...'text-lg leading-7 overview xl:w-3/4 href-text v-html text-danahergray-700'.split(' '));
