+++ blocks/product-parts/product-parts.js
@@ -4,7 +4,7 @@ import {
 import { getProductResponse } from '../../scripts/scripts.js';
 
 export default async function decorate(block) {
-  const response = getProductResponse();
+  const response = await getProductResponse();
   if (response?.length > 0 && response[0].raw?.bundlepreviewjson) {
     try {
       block.append(div(
